:art: #37 create GameCanvas
