+++ src/components/GameCanvas/index.scss
@@ -1,0 +1,7 @@
+.game-canvas {
+  position: absolute;
+  left: 0;
+  right: 0;
+  width: 100%;
+  height: 100%;
+}
